Switch to using a virtual base class instead of CRTP for shapes. Use raw pointers instead of shared_ptr for Intersection.
This was done due to the increasing complexity of the project. I also benchmarked both approaches and the performance difference was nearly non-exsistent. The reason why I thought that this was faster was because I used shared_ptr in intersection which made an allocation everytime an intersection was created. Lesson learn: do not just benchmark, but also profile to see where the time is going.